requirements.txt
requests
slack_sdk
beautifulsoup4
python-dotenv